Front-end traffic types
GET POINTS: added &layer=X parameter to points request. X is 1=cars
only, 2=pub only, 3=both (treat like a bitfield). The backend should
discriminate on the points returned if this is set. It should probably
also return this in the JSON so (if we had implemented) we can use a
different type of marker and display the tag.

CHECK IN: added &layer=X parameter to request. X is 1=car, 2=pub.